[conv]関数共通化、更新が無いファイルは処理しない

+++ tool/conv.ts
@@ -3,6 +3,7 @@ import * as path from 'path';
 // import marked from 'marked';
 import marked = require('marked');
 import imageSize from 'image-size';
+import * as base from './base-conv';
 
 const docsDirName = 'docs';
 const mdDirName = 'md';
@@ -15,30 +16,34 @@ function convMain(): void {
     console.log('run conv.ts');
 
     const isDeploy = process.argv.some((v) => v === '--deploy');
+    const isFullBuild = isDeploy || process.argv.some((v) => v === '--fullbuild');
     const curDir = process.cwd();
 
     // docsを全削除
-    try {
-        fs.rmSync(path.join(curDir, 'docs'), { recursive: true, force: true });
-    } catch (e) {
-        //fs.rmdirSync(path.join(curDir, 'docs'), { recursive: true });
+    if (isFullBuild) {
+        try {
+            // node v14
+            fs.rmSync(path.join(curDir, 'docs'), { recursive: true, force: true });
+        } catch (e) {
+            //fs.rmdirSync(path.join(curDir, 'docs'), { recursive: true });
+        }
     }
 
     const mdDir = path.join(curDir, mdDirName);
-    const mdPaths = lsrSync(mdDir);
+    const mdPaths = base.lsrSync(mdDir);
 
     const tmplPath = path.join(curDir, templateDirName, 'template.html');
     const tmplHtmlData = fs.readFileSync(tmplPath);
 
     // 共通部品md
     const sidebarPath = path.join(mdTemplateDirName, 'tp-sidebar.md');
-    const sidebarHtmlData = convMdToHtml(sidebarPath);
+    const sidebarHtmlData = base.convMdToHtml(sidebarPath);
     const navPcPath = path.join(mdTemplateDirName, 'tp-nav-pc.md');
-    const navPcHtmlData = convMdToHtml(navPcPath);
+    const navPcHtmlData = base.convMdToHtml(navPcPath);
     const navMobilePath = path.join(mdTemplateDirName, 'tp-nav-mobile.md');
-    const navMobileHtmlData = convMdToHtml(navMobilePath);
+    const navMobileHtmlData = base.convMdToHtml(navMobilePath);
     const footerPath = path.join(mdTemplateDirName, 'tp-footer.md');
-    const footerHtmlData = convMdToHtml(footerPath);
+    const footerHtmlData = base.convMdToHtml(footerPath);
 
     const dlDirPath = path.join(curDir, 'md', 'dl');
 
@@ -47,37 +52,38 @@ function convMain(): void {
             if (mdPath.match(/tp-.*.md$/)) {
                 return;
             }
-            // markdown -> html 変換
-            const mdData = fs.readFileSync(mdPath);
-            const htmlData = marked(mdData.toString());
+            let htmlPath = base.toHtmlPath(mdPath).replace(/index.html$/g, 'index0.html');
+            if (base.isUpdateFile(mdPath, htmlPath)) {
+                // markdown -> html 変換
+                const mdData = fs.readFileSync(mdPath);
+                const htmlData = marked(mdData.toString());
 
-            // テンプレートhtmlに、markdownのhtmlを埋め込み
-            const tmplData = Buffer.from(tmplHtmlData).toString();
-            const outData = tmplData.replace('%content', htmlData)
-                .replace(/%random/g, Math.floor(Math.random() * 1000000000).toString())
-                .replace(/%title/g, getTitle(htmlData))
-                .replace(/%sidebar/g, sidebarHtmlData)
-                .replace(/%nav_pc/g, navPcHtmlData)
-                .replace(/%nav_mobile/g, navMobileHtmlData)
-                .replace(/src=\"(.*?)\"/g,
-                    function (match, p1, offset, string) {
-                        // imgタグにwidth,heightを追加する
-                        const imgSize = imageSize(path.join(path.parse(mdPath).dir, p1));
-                        const width = imgSize ? 'width="' + imgSize.width + '"' : '';
-                        const height = imgSize ? 'height="' + imgSize.height + '"' : '';
-                        return 'src="' + p1 + '" ' + width + ' ' + height;
-                    })
-                .replace(/href="\.\/"/g, 'href="./index0.html"')
-                .replace(/href=\"(\.\/)*([a-zA-Z0-9-_]*)(.md|.html)*(#[a-zA-Z0-9-_]*)*\"/g,
-                    'href=\"$1$2.html$4"')
-                .replace(/%footer/g, footerHtmlData);
+                // テンプレートhtmlに、markdownのhtmlを埋め込み
+                const tmplData = Buffer.from(tmplHtmlData).toString();
+                const outData = tmplData.replace('%content', htmlData)
+                    .replace(/%random/g, Math.floor(Math.random() * 1000000000).toString())
+                    .replace(/%title/g, base.getTitle(htmlData))
+                    .replace(/%sidebar/g, sidebarHtmlData)
+                    .replace(/%nav_pc/g, navPcHtmlData)
+                    .replace(/%nav_mobile/g, navMobileHtmlData)
+                    .replace(/src=\"(.*?)\"/g,
+                        function (match, p1, offset, string) {
+                            // imgタグにwidth,heightを追加する
+                            const imgSize = imageSize(path.join(path.parse(mdPath).dir, p1));
+                            const width = imgSize ? 'width="' + imgSize.width + '"' : '';
+                            const height = imgSize ? 'height="' + imgSize.height + '"' : '';
+                            return 'src="' + p1 + '" ' + width + ' ' + height;
+                        })
+                    .replace(/href="\.\/"/g, 'href="./index0.html"')
+                    .replace(/href=\"(\.\/)*([a-zA-Z0-9-_]*)(.md|.html)*(#[a-zA-Z0-9-_]*)*\"/g,
+                        'href=\"$1$2.html$4"')
+                    .replace(/%footer/g, footerHtmlData);
 
-            let htmlPath = toHtmlPath(mdPath);
-            htmlPath = htmlPath.replace(/index.html$/g, 'index0.html');
-            fs.mkdirSync(path.dirname(htmlPath), { recursive: true });
-            fs.writeFileSync(htmlPath, outData);
+                fs.mkdirSync(path.dirname(htmlPath), { recursive: true });
+                fs.writeFileSync(htmlPath, outData);
 
-            console.log('convert md:', htmlPath);
+                console.log('convert md:', htmlPath);
+            }
         } else {
             // dlフォルダはデプロイ時のみコピー
             if (!isDeploy && mdPath.indexOf(dlDirPath) != -1) {
@@ -85,13 +91,13 @@ function convMain(): void {
             }
             // ファイルコピー(md -> docs)
             const relativePath = mdPath.replace(mdDir, '');
+            const inPath = path.join(mdDir, relativePath);
             const outPath = path.join(curDir, docsDirName, relativePath);
-            fs.mkdirSync(path.dirname(outPath), { recursive: true });
-            fs.copyFileSync(
-                path.join(mdDir, relativePath),
-                outPath
-            );
-            console.log('copy file :', outPath);
+            if (base.isUpdateFile(inPath, outPath)) {
+                fs.mkdirSync(path.dirname(outPath), { recursive: true });
+                fs.copyFileSync(inPath, outPath);
+                console.log('copy file :', outPath);
+            }
         }
     });
 
@@ -101,67 +107,5 @@ function convMain(): void {
         path.join(curDir, templateDirName, 'style.css'),
         outPath
     );
-}
-
-/**
- * mdファイルを読み込んで、変換、HTMLデータを返す
- * @param htmlPath 
- * @returns htmlのデータ
- */
-function convMdToHtml(htmlPath: string): string {
-    const mdData = fs.readFileSync(htmlPath);
-    const htmlData = marked(mdData.toString());
-    return htmlData;
-}
-
-/**
- * mdファイルパスをhtmlファイルパスにする
- * @param mdDir 
- * @param mdPath 
- * @returns 
- */
-function toHtmlPath(mdPath: string): string {
-    const relativePath = path.parse(mdPath).base
-    const tempPath = path.join(process.cwd(), 'docs', relativePath);
-    const htmlPath = tempPath.replace(/.md$/, '.html');
-    return htmlPath;
-}
-
-/**
- * ファイルの一覧を返す（サブフォルダ内のファイル含む）
- * @param dir 
- * @returns ファイルパス配列（絶対パス）
- */
-function lsrSync(dir: string): string[] {
-    const retAry: string[] = [];
-    const pathList = fs.readdirSync(dir);
-    pathList.map(p => path.join(dir, p))
-        .forEach(p => {
-            if (fs.statSync(p).isDirectory()) {
-                lsrSync(p).forEach(pp => {
-                    retAry.push(pp);
-                });
-            } else {
-                retAry.push(p);
-            }
-        });
-    return retAry;
-}
-
-/**
- * <head><title>に入れる文字列を取得する（<h1>～<h3>の文字列を取得）
- * @param htmlData 
- * @returns 
- */
-function getTitle(htmlData: string): string {
-    const start = htmlData.match(/<h[1-3]\s*.*?>/);
-    const end = htmlData.match(/<\/h[1-3]>/);
-    let title = '';
-    if (start != null && end != null) {
-        const startPos = start.index! + start[0].length;
-        const endPos = end.index;
-        title = htmlData.substring(startPos, endPos);
-    }
-    return title;
 }
